examples/rarebench/rarebench_requirements.txt
benchflow
openai
anthropic
google-generativeai